refactor: reorganize database
