+++ internal/infrastructure/db/postgres/postgres.go
@@ -1,4 +1,4 @@
-package db
+package postgres
 
 import (
 	"fmt"
